# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dastyare Social — CS",
    "version": "1.0.0",
    "description": "REST API for Dastyare Social — Creator Studio. Manage posts (threads, shorts, media) and stories. Supports JSON and multipart uploads, reactions, views, and likes. Authentication via Your API key. See /llms.txt for LLM-oriented documentation."
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateStoryBody": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "image",
              "video"
            ],
            "nullable": true
          },
          "views": {
            "type": "string",
            "nullable": true
          },
          "likes": {
            "type": "string",
            "nullable": true
          },
          "media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoryMediaItem"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StoryMediaItem"
                }
              }
            ]
          }
        }
      },
...
      "MediaUploadResponse": {},
...
      "patchStoriesSchema": {
        "type": "object"
      },
...
      "StoriesCountResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          }
        },
        "required": [
          "total"
        ]
      },
      "StoriesListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {}
          },
          "total": {
            "type": "number"
          },
          "hasMore": {
            "type": "boolean"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
...
      "StoriesResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/StoriesListResponse"
          },
          {
            "$ref": "#/components/schemas/StoriesCountResponse"
          }
        ]
      },
...
      "StoryItemSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "image",
              "video"
            ]
          },
          "views": {
            "type": "string"
          },
          "likes": {
            "type": "string"
          },
          "media": {},
          "createdAt": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "type",
          "views",
          "likes",
          "media",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "StoryMediaItem": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "thumbnail": {
            "type": "string"
          },
          "caption": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      },
...
    },
...
  },
  "paths": {
    "/upload": {
      "post": {
        "operationId": "post-upload",
        "summary": "Upload Media — Single File",
        "description": "Single-file upload endpoint (multipart/form-data with a `file` field). Server-side upload — buffers the file then uploads to S3. Returns media metadata. For real upload progress, use `/api/upload/presign` + `/api/upload/confirm` (direct browser-to-S3) instead.",
        "tags": [
          "Media"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "x-jsonschema-contentMediaType": "application/octet-stream"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "description": "Multipart form data containing a file upload.",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResponse"
                },
                "examples": {
                  "example": {
                    "value": {
                      "url": "https://cdn.example.com/media/image/abc.jpg",
                      "key": "media/image/abc.jpg",
                      "kind": "image",
                      "mimeType": "image/jpeg",
                      "size": 12345,
                      "width": 1080,
                      "height": 1920,
                      "duration": 0,
                      "filename": "photo.jpg"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/upload-stream": {
      "post": {
        "operationId": "post-upload-stream",
        "summary": "Upload Media — Streaming with Progress",
        "description": "Upload a single media file with server-side progress tracking.",
        "tags": [
          "Media"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "x-jsonschema-contentMediaType": "application/octet-stream"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "description": "Multipart form data containing a file upload.",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/200 - SSE stream with progress/done/error events"
                }
              }
            }
          },
          "400 - Invalid request": {
            "$ref": "#/components/responses/400 - Invalid request"
          },
          "403 - Demo mode active": {
            "$ref": "#/components/responses/403 - Demo mode active"
          },
          "500 - Server error": {
            "$ref": "#/components/responses/500 - Server error"
          }
        }
      }
    },
...
    "/stories": {
      "get": {
        "operationId": "get-stories",
        "summary": "List Stories — Paginated Feed",
        "description": "Returns paginated stories. Use query type=count for total. Filter by kind=image or kind=video.",
        "tags": [
          "Stories"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "example"
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": 1
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "example"
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "example"
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "example"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoriesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-stories",
        "summary": "Create Story — Upload Media Via URL",
        "description": "Create a story from a JSON body. Media must be referenced by URL — upload files to `/api/upload` or `/api/upload-stream` first, then map the returned `url`, `kind` (as `type`), `width`, `height`, and `duration` into the `media` field below. Types: image, video. An array of media items creates one story per item.",
        "tags": [
          "Stories"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoryBody"
              },
              "examples": {
                "example": {
                  "value": {
                    "type": "image",
                    "media": {
                      "url": "https://cdn.example.com/media/image/abc.jpg",
                      "width": 1080,
                      "height": 1920
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryItemSchema"
                }
              }
            }
          },
          "StorySuccessResponse": {
            "$ref": "#/components/responses/StorySuccessResponse"
          }
        }
      }
    },
    "/stories/{story_id}": {
      "get": {
        "operationId": "get-stories-{story_id}",
        "summary": "Read Story — Get by ID",
        "description": "Returns a single story.",
        "tags": [
          "Stories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/200 StoryItemSchema"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-stories-{story_id}",
        "summary": "Story Actions — View or Like",
        "description": "Perform actions on a story. action=view increments views. action=like with direction=inc or dec toggles likes.",
        "tags": [
          "Stories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/201 StoryItemSchema"
                }
              }
            }
          },
          "200 StorySuccessResponse": {
            "$ref": "#/components/responses/200 StorySuccessResponse"
          }
        }
      },
      "patch": {
        "operationId": "patch-stories-{story_id}",
        "summary": "Update Story — Edit Fields",
        "description": "Partial update of story fields (type, views, likes, media).",
        "tags": [
          "Stories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "123"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/patchStoriesSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/200 StoryItemSchema"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-stories-{story_id}",
        "summary": "Delete Story — Remove by ID",
        "description": "Permanently deletes a story by ID.",
        "tags": [
          "Stories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "z.boolean() })",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/z.object({ success"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "API server"
    }
  ]
}